Moved libtx to core and changed all apparent references
