+++ chain/tests/data_file_integrity.rs
@@ -35,7 +35,7 @@ use core::global::{self, ChainTypes};
 use core::pow::{self, Difficulty};
 use core::{consensus, genesis};
 use keychain::{ExtKeychain, ExtKeychainPath, Keychain};
-use wallet::libtx;
+use core::libtx;
 
 fn clean_output_dir(dir_name: &str) {
 	let _ = fs::remove_dir_all(dir_name);
