+++ chain/tests/test_coinbase_maturity.rs
@@ -35,7 +35,7 @@ use core::global::{self, ChainTypes};
 use core::pow::Difficulty;
 use core::{consensus, pow};
 use keychain::{ExtKeychain, ExtKeychainPath, Keychain};
-use wallet::libtx::{self, build};
+use core::libtx::{self, build};
 
 fn clean_output_dir(dir_name: &str) {
 	let _ = fs::remove_dir_all(dir_name);
