+++ core/fuzz/src/main.rs
@@ -7,8 +7,8 @@ use grin_core::core::{Block, BlockHeader, CompactBlock, Transaction};
 use grin_core::ser;
 use grin_keychain::keychain::ExtKeychain;
 use grin_keychain::Keychain;
-use grin_wallet::libtx::build::{input, output, transaction, with_fee};
-use grin_wallet::libtx::reward;
+use grin_core::libtx::build::{input, output, transaction, with_fee};
+use grin_core::libtx::reward;
 use std::fs::{self, File};
 use std::path::Path;
