+++ core/src/libtx/proof.rs
