+++ core/src/libtx/slate.rs
@@ -19,10 +19,10 @@ use rand::thread_rng;
 use std::sync::Arc;
 use uuid::Uuid;
 
-use core::core::committed::Committed;
-use core::core::transaction::kernel_sig_msg;
-use core::core::verifier_cache::LruVerifierCache;
-use core::core::{amount_to_hr_string, Transaction};
+use core::committed::Committed;
+use core::transaction::kernel_sig_msg;
+use core::verifier_cache::LruVerifierCache;
+use core::{amount_to_hr_string, Transaction};
 use keychain::{BlindSum, BlindingFactor, Keychain};
 use libtx::error::{Error, ErrorKind};
 use libtx::{aggsig, build, tx_fee};
