+++ core/tests/block.rs
@@ -16,7 +16,6 @@ extern crate chrono;
 extern crate grin_core;
 extern crate grin_keychain as keychain;
 extern crate grin_util as util;
-extern crate grin_wallet as wallet;
 
 use std::sync::Arc;
 use std::time::Instant;
@@ -34,9 +33,9 @@ use grin_core::core::verifier_cache::{LruVerifierCache, VerifierCache};
 use grin_core::core::Committed;
 use grin_core::core::{Block, BlockHeader, CompactBlock, KernelFeatures, OutputFeatures};
 use grin_core::{global, ser};
+use grin_core::libtx::build::{self, input, output, with_fee};
 use keychain::{BlindingFactor, ExtKeychain, Keychain};
 use util::secp;
-use wallet::libtx::build::{self, input, output, with_fee};
 
 fn verifier_cache() -> Arc<RwLock<VerifierCache>> {
 	Arc::new(RwLock::new(LruVerifierCache::new()))
