+++ core/tests/common/mod.rs
@@ -17,14 +17,13 @@
 extern crate grin_core;
 extern crate grin_keychain as keychain;
 extern crate grin_util as util;
-extern crate grin_wallet as wallet;
 
 use grin_core::core::block::{Block, BlockHeader};
 use grin_core::core::Transaction;
 use grin_core::pow::Difficulty;
 use keychain::{Identifier, Keychain};
-use wallet::libtx::build::{self, input, output, with_fee};
-use wallet::libtx::reward;
+use grin_core::libtx::build::{self, input, output, with_fee};
+use grin_core::libtx::reward;
 
 // utility producing a transaction with 2 inputs and a single outputs
 pub fn tx2i1o() -> Transaction {
