+++ pool/tests/transaction_pool.rs
@@ -33,7 +33,7 @@ use core::core::verifier_cache::LruVerifierCache;
 use core::core::{transaction, Block, BlockHeader};
 use core::pow::Difficulty;
 use keychain::{ExtKeychain, Keychain};
-use wallet::libtx;
+use core::libtx;
 
 /// Test we can add some txs to the pool (both stempool and txpool).
 #[test]
