+++ wallet/tests/common/testclient.rs
@@ -42,7 +42,7 @@ use core::{pow, ser};
 use keychain::Keychain;
 
 use util::secp::pedersen;
-use wallet::libtx::slate::Slate;
+use core::libtx::slate::Slate;
 use wallet::libwallet::types::*;
 use wallet::{controller, libwallet, WalletCommAdapter, WalletConfig};
